@@ -1,7 +1,7 @@
 package com.hencoder.hencoderpracticedraw6.practice;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.Button;
@@ -9,10 +9,13 @@ import android.widget.ImageView;
 import android.widget.RelativeLayout;
 
 import com.hencoder.hencoderpracticedraw6.R;
+import com.hencoder.hencoderpracticedraw6.Utils;
 
 public class Practice02Rotation extends RelativeLayout {
     Button animateBt;
     ImageView imageView;
+    private int translateState = 0;
+    private static final Integer translateStateCap = 5;
 
     public Practice02Rotation(Context context) {
         super(context);
@@ -36,7 +39,32 @@ public class Practice02Rotation extends RelativeLayout {
         animateBt.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(final View v) {
-                // // TODO 在这里处理点击事件，通过 View.animate().rotation/X/Y() 来让 View 旋转
+                switch (translateState) {
+                    case 0:
+                        imageView.animate().rotation(180);
+                        break;
+                    case 1:
+                        imageView.animate().rotation(0);
+                        break;
+                    case 2:
+                        // Use x as axis to rotate
+                        imageView.animate().rotationX(180);
+                        break;
+                    case 3:
+                        imageView.animate().rotationX(0);
+                        break;
+                    case 4:
+                        imageView.animate().rotationY(180);
+                        break;
+                    case 5:
+                        imageView.animate().rotationY(0);
+                        break;
+                }
+
+                translateState++;
+                if (translateState == translateStateCap) {
+                    translateState = 0;
+                }
             }
         });
     }
