@@ -1,7 +1,7 @@
 package com.hencoder.hencoderpracticedraw6.practice;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.Button;
@@ -13,6 +13,8 @@ import com.hencoder.hencoderpracticedraw6.R;
 public class Practice04Alpha extends RelativeLayout {
     Button animateBt;
     ImageView imageView;
+    private int translateState = 0;
+    private static final Integer translateStateCap = 2;
 
     public Practice04Alpha(Context context) {
         super(context);
@@ -36,7 +38,20 @@ public class Practice04Alpha extends RelativeLayout {
         animateBt.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(final View v) {
-                // TODO 在这里处理点击事件，通过 View.animate().alpha() 来改变 View 的透明度
+                switch (translateState) {
+                    case 0:
+                        // alpha 0 is transparent
+                        imageView.animate().alpha(0);
+                        break;
+                    case 1:
+                        imageView.animate().alpha(1);
+                        break;
+                }
+
+                translateState++;
+                if (translateState == translateStateCap) {
+                    translateState = 0;
+                }
             }
         });
     }
