@@ -1,18 +1,22 @@
 package com.hencoder.hencoderpracticedraw6.practice;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.constraint.ConstraintLayout;
+import androidx.annotation.Nullable;
+import androidx.constraintlayout.widget.ConstraintLayout;
+
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
 
 import com.hencoder.hencoderpracticedraw6.R;
+import com.hencoder.hencoderpracticedraw6.Utils;
 
 public class Practice05MultiProperties extends ConstraintLayout {
     Button animateBt;
     ImageView imageView;
+    private int translateState = 0;
+    private static final Integer translateStateCap = 2;
 
     public Practice05MultiProperties(Context context) {
         super(context);
@@ -37,8 +41,30 @@ public class Practice05MultiProperties extends ConstraintLayout {
         imageView.setAlpha(0f);
         animateBt.setOnClickListener(new OnClickListener() {
             @Override
-            public void onClick(View v) {
-                // TODO 在这里处理点击事件，同时对多个属性做动画
+            public void onClick(final View v) {
+                switch (translateState) {
+                    case 0:
+                        imageView.animate()
+                            .translationX(Utils.dpToPixel(200))
+                            .scaleX(1)
+                            .scaleY(1)
+                            .rotation(360)
+                            .alpha(1);
+                        break;
+                    case 1:
+                        imageView.animate()
+                            .translationX(0)
+                            .scaleX(0)
+                            .scaleY(0)
+                            .rotation(0)
+                            .alpha(0);
+                        break;
+                }
+
+                translateState++;
+                if (translateState == translateStateCap) {
+                    translateState = 0;
+                }
             }
         });
     }
