@@ -1,7 +1,7 @@
 package com.hencoder.hencoderpracticedraw6.sample;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.Button;
