@@ -1,8 +1,9 @@
 package com.hencoder.hencoderpracticedraw6.sample;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.constraint.ConstraintLayout;
+import androidx.annotation.Nullable;
+import androidx.constraintlayout.widget.ConstraintLayout;
+
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.Button;
