@@ -1,21 +1,25 @@
 package com.hencoder.hencoderpracticedraw6;
 
 import android.os.Bundle;
-import android.support.annotation.LayoutRes;
-import android.support.annotation.StringRes;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentPagerAdapter;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
+
+import androidx.annotation.LayoutRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.StringRes;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.viewpager2.adapter.FragmentStateAdapter;
+import androidx.viewpager2.widget.ViewPager2;
+
+import com.google.android.material.tabs.TabLayout;
+import com.google.android.material.tabs.TabLayoutMediator;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
     TabLayout tabLayout;
-    ViewPager pager;
+    ViewPager2 pager;
     List<PageModel> pageModels = new ArrayList<>();
 
     {
@@ -34,28 +38,29 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        pager = (ViewPager) findViewById(R.id.pager);
-        pager.setAdapter(new FragmentPagerAdapter(getSupportFragmentManager()) {
+        pager = (ViewPager2) findViewById(R.id.pager);
+        pager.setAdapter(new FragmentStateAdapter(this) {
 
             @Override
-            public Fragment getItem(int position) {
-                PageModel pageModel = pageModels.get(position);
-                return PageFragment.newInstance(pageModel.sampleLayoutRes, pageModel.practiceLayoutRes);
-            }
-
-            @Override
-            public int getCount() {
+            public int getItemCount() {
                 return pageModels.size();
             }
 
+            @NonNull
             @Override
-            public CharSequence getPageTitle(int position) {
-                return getString(pageModels.get(position).titleRes);
+            public Fragment createFragment(int position) {
+                PageModel pageModel = pageModels.get(position);
+                return PageFragment.newInstance(pageModel.sampleLayoutRes, pageModel.practiceLayoutRes);
             }
         });
 
         tabLayout = (TabLayout) findViewById(R.id.tabLayout);
-        tabLayout.setupWithViewPager(pager);
+        new TabLayoutMediator(tabLayout, pager, new TabLayoutMediator.TabConfigurationStrategy() {
+            @Override
+            public void onConfigureTab(@NonNull TabLayout.Tab tab, int position) {
+                tab.setText(pageModels.get(position).titleRes);
+            }
+        }).attach();
     }
 
     @Override
@@ -64,8 +69,10 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private class PageModel {
-        @LayoutRes int sampleLayoutRes;
-        @StringRes int titleRes;
+        @LayoutRes
+        int sampleLayoutRes;
+        @StringRes
+        int titleRes;
         @LayoutRes int practiceLayoutRes;
 
         PageModel(@LayoutRes int sampleLayoutRes, @StringRes int titleRes, @LayoutRes int practiceLayoutRes) {
